[QUAD-210] Making code changes to get list of projects using a given resolver and adding delete icon instead of delete button

+++ Quadriga/src/main/java/edu/asu/spring/quadriga/service/resolver/impl/ProjectHandleResolverManager.java
@@ -89,6 +89,11 @@ public class ProjectHandleResolverManager implements IProjectHandleResolverManag
     @Transactional
     public boolean deleteProjectHandleResolver(IProjectHandleResolver resolver) {
 
+        if (resolverDao.getProjectsForResolverId(resolver.getId()).size() > 0) {
+            throw new RuntimeException("resolver cannot be deleted as it is being used by project");
+        }
+        ;
+
         ProjectHandleResolverDTO resolverDto = mapper.mapProjectHandleResolver(resolver);
         resolverDao.deleteDTO(resolverDto);
